handled merge conflict with branch #20
